Take tests out of classes
