+++ src/edu/umass/cs/utils/Util.java
@@ -896,44 +896,4 @@ public class Util {
 		return null;
 	}
 
-	public static class UtilTest extends DefaultTest {
-		/**
-		 * @throws JSONException
-		 */
-		@SuppressWarnings("unchecked")
-		@Test
-		public void test_01_JSONObjectToMap() throws JSONException {
-			Util.assertAssertionsEnabled();
-			Map<String, ?> map = (JSONObjectToMap(new JSONObject()
-					.put("hello", "world")
-					.put("collField", Arrays.asList("hello", "world", 123))
-					.put("jsonField", new JSONObject().put("foo", true))));
-			System.out.println(map);
-			org.junit.Assert.assertTrue(map.get("jsonField") instanceof Map
-					&& (Boolean) ((Map<String, ?>) map.get("jsonField"))
-							.get("foo"));
-
-		}
-
-		/**
-		 * @throws JSONException
-		 */
-		@SuppressWarnings("unchecked")
-		@Test
-		public void test_01_JSONArrayToMap() throws JSONException {
-			Util.assertAssertionsEnabled();
-			List<?> list = (JSONArrayToList(new JSONArray()
-
-			.put("hello") // 0
-
-					.put(Arrays.asList("hello", "world", 123)) // 1
-
-					.put(new JSONObject().put("foo", true)))); // 2
-
-			System.out.println(list);
-			org.junit.Assert.assertTrue(list.get(2) instanceof Map
-					&& (Boolean) ((Map<String, ?>) list.get(2)).get("foo"));
-
-		}
-	}
 }
